+++ endesive/__init__.py
@@ -2,6 +2,6 @@
 
 __author__ = 'Grzegorz Makarewicz'
 __license__ = 'MIT'
-__version__ = '1.5.12'
+__version__ = '2.0.0'
 
 __all__ = [__author__, __license__, __version__]