+++ examples/pdf-sign-cms-m32.py
@@ -15,15 +15,15 @@ def main():
     date = datetime.datetime.utcnow() - datetime.timedelta(hours=12)
     date = date.strftime('%Y%m%d%H%M%S+00\'00\'')
     dct = {
-        b'sigflags': 3,
-        b'sigpage': 0,
-        b'sigbutton': True,
-        b'contact': b'mak@trisoft.com.pl',
-        b'location': b'Szczecin',
-        b'signingdate': date.encode(),
-        b'reason': b'Dokument podpisany cyfrowo',
-        b'signature': b'Dokument podpisany cyfrowo',
-        b'signaturebox': (0, 0, 100, 100),
+        'sigflags': 3,
+        'sigpage': 0,
+        'sigbutton': True,
+        'contact': 'mak@trisoft.com.pl',
+        'location': 'Szczecin',
+        'signingdate': date.encode(),
+        'reason': 'Dokument podpisany cyfrowo',
+        'signature': 'Dokument podpisany cyfrowo',
+        'signaturebox': (0, 0, 100, 100),
     }
     with open('/home/mak/Dokumenty/m32/ssl/cert.p12', 'rb') as fp:
         p12 = pkcs12.load_key_and_certificates(fp.read(), b'1234', backends.default_backend())